Add files via upload
The dataset we have contains search queries that lead users to a specific website, along with associated metrics. The columns in this dataset are:  

Top Queries: The actual search terms used by users.  
Clicks: The number of times users clicked on the website after using the query.  
Impressions: The number of times the website appeared in search results for the query.  
CTR (Click Through Rate): The ratio of clicks to impressions, indicating the effectiveness of the query in leading users to the website.  
Position: The average ranking of the website in search results for the query.